Add common check_other_party validation for mutations
